requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy<2
python-multipart